Adding WHO files
The who_data notebook takes all the csvs located in /data and merges into one flat dataframe. I think it's in "First Normal Form". Most of the data is from 2016 but a few files had longitudinal data from mulitple years, so there are a lot rows with mostly missing data. We can clean it up more if need be.